+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
-**UNRELEASED**
+**4.0.0rc4** (2023-04-15)
 
 - Fixed imports guarded by ``if TYPE_CHECKING:`` when used with subscripts
   (``SomeType[...]``) being replaced with ``Any[...]`` instead of just ``Any``